+++ ntex-grpc-codegen/src/config.rs
@@ -21,14 +21,9 @@ impl NGrpcGenerator {
         Self { pconfig: cfg }
     }
 
-    /// Map protobuf bytes type to custom rust type
-    pub fn map_bytes(&mut self, path: &str, rust_type: &str) {
-        let _ = self.pconfig.bytes(&[path], rust_type);
-    }
-
-    /// Map protobuf string type to custom rust type
-    pub fn map_string(&mut self, path: &str, rust_type: &str) {
-        let _ = self.pconfig.string(&[path], rust_type);
+    /// Map protobuf type to custom rust type
+    pub fn map_field_type(&mut self, path: &str, rust_type: &str) {
+        let _ = self.pconfig.map_field_type(&[path], rust_type);
     }
 
     /// Compile `.proto` files into Rust files during a Cargo build with additional code generator