+++ ntex-grpc-codegen/CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+## [0.1.5] - 2022-07-08
+
+* Allow to replace any protobuf field with custom rust type
+
 ## [0.1.4] - 2022-07-06
 
 * Generate service defs manually